+++ lib/beyond_canvas.rb
@@ -1,20 +1,14 @@
-require "beyond_canvas/version"
+# frozen_string_literal: true
 
-require "bourbon"
-require "slim-rails"
-require "inline_svg"
+require 'beyond_canvas/engine'
+
+require 'colorize'
+
+require 'bourbon'
+require 'slim-rails'
+require 'inline_svg'
 
 module BeyondCanvas
-  class Error < StandardError; end
-
-  class Engine < ::Rails::Engine
-    config.before_initialize do
-      if config.action_view.javascript_expansions
-        config.action_view.javascript_expansions[:beyond_canvas] = %w(beyond_canvas)
-      end
-    end
-  end
-
   class << self
     attr_accessor :configuration
   end