+++ lib/generators/beyond_canvas/custom_styles_generator.rb
@@ -3,10 +3,10 @@
 module BeyondCanvas
   module Generators
     class CustomStylesGenerator < Rails::Generators::Base
-      source_root File.expand_path("../../templates", __FILE__)
+      source_root File.expand_path('../templates', __dir__)
 
       def copy_initializer
-        template "beyond_canvas_custom_styles.sass", "app/assets/stylesheets/_beyond_canvas_custom_styles.sass"
+        template 'beyond_canvas_custom_styles.sass', 'app/assets/stylesheets/_beyond_canvas_custom_styles.sass'
       end
     end
   end