+++ lib/generators/beyond_canvas/install_generator.rb
@@ -3,11 +3,10 @@
 module BeyondCanvas
   module Generators
     class InstallGenerator < Rails::Generators::Base
-      source_root File.expand_path("../../templates", __FILE__)
+      source_root File.expand_path('../templates', __dir__)
 
       def copy_initializer
-        template "beyond_canvas.rb", "config/initializers/beyond_canvas.rb"
-        template "beyond_canvas_form_utils.rb", "config/initializers/beyond_canvas_form_utils.rb"
+        template 'beyond_canvas.rb', 'config/initializers/beyond_canvas.rb'
       end
     end
   end